Move records to new files
